Example for verification how often methods of an mockobject are called

+++ src/Library/Book.java
@@ -29,7 +29,7 @@ public class Book{
         return chapterNumber;
     }
 
-    public float averagePagesPerChapter(){
+    public double averagePagesPerChapter(){
         return pageNumber/chapterNumber;
     }
 }
